+++ src/lib.rs
@@ -50,7 +50,7 @@ pub mod args {
 pub mod migration {
     #[cfg(feature = "dev")]
     pub use crate::hash::dev::hash_schema;
-    pub use crate::migrate::{Config, Migrated, Migrator, TransactionMigrate};
+    pub use crate::migrate::{Config, Migrated, Migrator, Synchronous, TransactionMigrate};
     pub use rust_query_macros::schema;
 }